profileSeachView DashBoardView のコンポーネント分割
