+++ frontend/src/components/profile/SearchHeader.vue
@@ -1,0 +1,34 @@
+<template>
+  <div class="search-header">
+    <h1 class="search-title">プロフィール検索</h1>
+    <p class="search-subtitle">
+      同じUSJ好きの仲間を見つけよう！
+    </p>
+  </div>
+</template>
+
+<script setup lang="ts">
+// プロップスは不要な単純なヘッダーコンポーネント
+</script>
+
+<style scoped>
+.search-header {
+  text-align: center;
+  margin-bottom: 2rem;
+  padding-bottom: 1rem;
+  border-bottom: 2px solid #f3f4f6;
+}
+
+.search-title {
+  font-size: 1.875rem;
+  font-weight: 700;
+  color: #1f2937;
+  margin: 0 0 0.5rem 0;
+}
+
+.search-subtitle {
+  color: #6b7280;
+  font-size: 1.1rem;
+  margin: 0;
+}
+</style>
