+++ frontend/src/components/profile/index.ts
@@ -2,4 +2,9 @@ export { default as ProfileHeader } from './ProfileHeader.vue'
 export { default as ProfileDisplay } from './ProfileDisplay.vue'
 export { default as ProfileForm } from './ProfileForm.vue'
 export { default as AttractionsSelector } from './AttractionsSelector.vue'
-export { default as HobbiesSelector } from './HobbiesSelector.vue' 
+export { default as HobbiesSelector } from './HobbiesSelector.vue'
+export { default as SearchHeader } from './SearchHeader.vue'
+export { default as SearchForm } from './SearchForm.vue'
+export { default as QuickSearch } from './QuickSearch.vue'
+export { default as SearchResults } from './SearchResults.vue'
+export { default as ProfileCard } from './ProfileCard.vue'
